Improve jar pre-caching
The old solution of overrding $HOME to /tmp and putting the caches there
doesn't work properly under apptainer, as it forcibly sets $HOME to the calling
user's $HOME and doesn't offer any way to disable this. It was also a problem
for apptainer configurations that bind /tmp into the container by default
(although at least that can be disabled with apptainer command line flags).
This change moves the caches to /opt/scyjava. The cache pre-warming has gotten
complex enough that it has been moved to a standalone python script that is
called during the docker build process.